added more styling
